refactor(utils): move sorting utils to a new file
